+++ tests/unit/pixel/image/GdTest.php
@@ -105,4 +105,19 @@ class GdTest extends TestCase
 		$this->assertSame('#000000', $colors[0]->toHexString());
 		$this->assertSame('#FFFFFF', $colors[1]->toHexString());
 	}
+
+	/**
+	 *
+	 */
+	public function testToDataUri(): void
+	{
+		$image = new Gd(__DIR__ . '/fixtures/onebyone.jpg');
+
+		$this->assertStringStartsWith('data:image/jpeg;base64,', $image->toDataUri());
+		$this->assertStringStartsWith('data:image/jpeg;base64,', $image->toDataUri('jpg'));
+		$this->assertStringStartsWith('data:image/jpeg;base64,', $image->toDataUri('jpeg'));
+		$this->assertStringStartsWith('data:image/jpeg;base64,', $image->toDataUri('image/jpg'));
+		$this->assertStringStartsWith('data:image/jpeg;base64,', $image->toDataUri('image/jpeg'));
+		$this->assertStringStartsWith('data:image/gif;base64,', $image->toDataUri('gif'));
+	}
 }